refactor(store): move create_n0_bao test util into util::tests::create_n0_bao
