+++ src/api/remote.rs
@@ -1073,10 +1073,11 @@ mod tests {
         protocol::{ChunkRangesExt, ChunkRangesSeq, GetRequest},
         store::{
             fs::{
-                tests::{create_n0_bao, test_data, INTERESTING_SIZES},
+                tests::{test_data, INTERESTING_SIZES},
                 FsStore,
             },
             mem::MemStore,
+            util::tests::create_n0_bao,
         },
         tests::{add_test_hash_seq, add_test_hash_seq_incomplete},
     };
